Refactor styles element
