+++ packages/server/src/urls.ts
@@ -1,0 +1,3 @@
+import { env } from "./env/server";
+
+export const buildCdnUrl = (src: string) => `${env.CDN_URL}/${src}`;
